rename library to ngx-drag-scroll
Closes #40
