+++ angular-cli.json
@@ -1,7 +1,7 @@
 {
   "project": {
     "version": "1.0.0-beta.28.3",
-    "name": "angular2-drag-scroll"
+    "name": "ngx-drag-scroll"
   },
   "apps": [
     {
@@ -35,7 +35,7 @@
       "files": "src/**/*.ts",
       "exclude": [
         "src/interface/drag-scroll-option.d.ts",
-        "src/angular2-drag-scroll.d.ts"
+        "src/ngx-drag-scroll.d.ts"
       ],
       "project": "demo/tsconfig.json"
     }
