+++ demo/app/app.module.ts
@@ -9,13 +9,14 @@ import 'style-loader!@angular/material/prebuilt-themes/indigo-pink.css';
 
 import { AppComponent } from './app.component';
 import { HeaderComponent } from './header/header.component';
-import { DragScrollModule } from '../../src/angular2-drag-scroll';
+import { DragScrollModule } from '../../src/ngx-drag-scroll';
 import { NotFoundComponent } from './not-found/not-found.component';
 import { AppRoutingModule } from './app-routing.module';
 import { HomeComponent } from './home/home.component';
 import { FooterComponent } from './footer/footer.component';
 import { GithubComponent } from './github/github.component';
 import { FlexLayoutModule } from '@angular/flex-layout';
+import { TestComponent } from './ngx-drag-scroll.spec';
 
 
 
@@ -26,6 +27,7 @@ import { FlexLayoutModule } from '@angular/flex-layout';
     NotFoundComponent,
     FooterComponent,
     GithubComponent,
+    TestComponent,
     HomeComponent
   ],
   imports: [
