+++ demo/app/footer/footer.component.html
@@ -2,8 +2,8 @@
   Hand crafted with love by <!-- /react-text -->
   <a href="https://github.com/bfwg" style="color: rgba(255, 255, 255, 0.870588);">Fan Jin</a>
   <!-- react-text: 105 --> and our awesome <!-- /react-text -->
-  <a href="https://github.com/bfwg/angular2-drag-scroll/graphs/contributors" style="color: rgba(255, 255, 255, 0.870588);">contributors</a><!-- react-text: 107 -->.<!-- /react-text -->
+  <a href="https://github.com/bfwg/ngx-drag-scroll/graphs/contributors" style="color: rgba(255, 255, 255, 0.870588);">contributors</a><!-- react-text: 107 -->.<!-- /react-text -->
 </p>
-<a style="margin-top: 22px;" mat-icon-button href="https://github.com/bfwg/angular2-drag-scroll">
+<a style="margin-top: 22px;" mat-icon-button href="https://github.com/bfwg/ngx-drag-scroll">
   <img src="assets/img/github.png"/>
 </a>
