+++ demo/app/github/github.component.html
@@ -1,4 +1,4 @@
 <h3>Want to help make this project awesome? Check out our repo.</h3>
-<a href="https://github.com/bfwg/angular2-drag-scroll" mat-raised-button matRipple>
+<a href="https://github.com/bfwg/ngx-drag-scroll" mat-raised-button matRipple>
   <span>GITHUB</span>
 </a>
