+++ demo/app/header/header.component.html
@@ -1,6 +1,6 @@
 <mat-toolbar color="primary" class="app-navbar">
   <button mat-button matRipple routerLink="/">
     <img alt="Angular" class="app-angular-logo" src="assets/img/angular-white-transparent.svg">
-    <span>Angular2-drag-scroll</span>
+    <span>ngx-drag-scroll</span>
   </button>
 </mat-toolbar>
