+++ demo/app/home/home.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, ElementRef, Renderer2, ViewChild } from '@angular/core';
 import { DomSanitizer } from '@angular/platform-browser';
 import { MatIconRegistry } from '@angular/material';
-import { DragScroll } from '../../../src/angular2-drag-scroll';
+import { DragScroll } from '../../../src/ngx-drag-scroll';
 
 @Component({
   selector: 'app-home',
