+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "angular2-drag-scroll",
+  "name": "ngx-drag-scroll",
   "version": "1.5.1",
-  "description": "Lightweight drag to scroll directive for Angular2",
+  "description": "Lightweight drag to scroll directive for Angular2+",
   "contributors": [
     {
       "name": "Fan Jin",
@@ -23,7 +23,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/bfwg/angular2-drag-scroll.git"
+    "url": "git+https://github.com/bfwg/ngx-drag-scroll.git"
   },
   "keywords": [
     "Angular 2",
@@ -33,9 +33,9 @@
   "author": "Fan Jin",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/bfwg/angular2-drag-scroll.git"
+    "url": "https://github.com/bfwg/ngx-drag-scroll.git"
   },
-  "homepage": "https://github.com/bfwg/angular2-drag-scroll#readme",
+  "homepage": "https://github.com/bfwg/ngx-drag-scroll#readme",
   "peerDependencies": {
     "@angular/common": "^5.0.0",
     "@angular/compiler": "^5.0.0",
